+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 * Testing
+    * Pages access tests.
     * User factory.
 * Register validation
     * Promotion must now be within an acceptable range.